pyproject.toml
[project]
name = "graphiti-mcp-server"
version = "0.1.0"
description = "Template for mcp server with connection to database"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi==0.116.1",
    "mcp==1.12.4",
    "pandas==2.3.1",
    "pydantic==2.11.7",
    "asyncpg>=0.30.0",
    "pytest>=8.4.1",
    "fastmcp>=2.11.3",
    "graphiti-core>=0.25.5",
    "python-dotenv>=1.1.1",
]
